+++ frontend/src/services/followService.js
@@ -3,10 +3,18 @@ import API_ROUTES from "../constants/apiRoutes";
 
 const followService = {
   followUser: async (userId) => {
+    if (!userId) {
+      return;
+    }
+
     await mainApi.post(`${API_ROUTES.API.FOLLOWS}/${userId}`, {});
   },
 
   unfollowUser: async (userId) => {
+    if (!userId) {
+      return;
+    }
+
     await mainApi.delete(`${API_ROUTES.API.FOLLOWS}/${userId}`);
   },
 };
