+++ frontend/src/services/postService.js
@@ -3,18 +3,30 @@ import API_ROUTES from "../constants/apiRoutes";
 
 const postService = {
   getPost: async (postId) => {
+    if (!postId) {
+      return;
+    }
+
     const result = await mainApi.get(`${API_ROUTES.API.POSTS}/${postId}`);
 
     return result.data;
   },
 
   addPost: async (post) => {
+    if (!post) {
+      return;
+    }
+
     const result = await mainApi.post(`${API_ROUTES.API.POSTS}`, post);
 
     return result.data;
   },
 
   getPostComments: async (postId, offset) => {
+    if (!postId) {
+      return;
+    }
+
     const result = await mainApi.get(
       `${API_ROUTES.API.POSTS}/${postId}/comments?offset=${offset}`
     );
@@ -23,6 +35,10 @@ const postService = {
   },
 
   addPostComment: async ({ postId, content, parentId = null }) => {
+    if (!postId) {
+      return;
+    }
+
     const result = await mainApi.post(
       `${API_ROUTES.API.POSTS}/${postId}/comments`,
       {
@@ -35,6 +51,10 @@ const postService = {
   },
 
   getPostLikeDetails: async (postId) => {
+    if (!postId) {
+      return;
+    }
+
     const results = await mainApi.get(
       `${API_ROUTES.API.POSTS}/${postId}/likes_details`
     );
@@ -43,14 +63,26 @@ const postService = {
   },
 
   addPostLike: async (postId) => {
+    if (!postId) {
+      return;
+    }
+
     await mainApi.post(`${API_ROUTES.API.POSTS}/${postId}/likes`, {});
   },
 
   deletePost: async (postId) => {
+    if (!postId) {
+      return;
+    }
+
     await mainApi.delete(`${API_ROUTES.API.POSTS}/${postId}`);
   },
 
   deletePostLike: async (postId) => {
+    if (!postId) {
+      return;
+    }
+
     await mainApi.delete(`${API_ROUTES.API.POSTS}/${postId}/likes`);
   },
 };
